refactor: update import paths to streamline type imports and adjust styles file extension
